Serve HTML static files on explicit ".html" requests (#110)
* Remove staticFilePath check

Removing staticFilePath check, as vite.middlewares are now mounted before the injectViteIndexMiddleware function. This adjustment is made because the staticFilePath check doesn't provide any additional benefits, aside from blocking requests like /index.html on the development server.

* Serve */index.html using IndexMiddlewares

Allowing requests for html files to be transformable. This is achieved by skipping files with .htm or .html in the Static Middleware, allowing them to be handled by the IndexMiddlewares.

* Correction

* Return closest index file if no html file found

Previously, my implementation involved checking if a file ended with the html extension and then verifying its existence. If the file didn't exist, it would trigger next(), eventually resulting in a 404 error. However, I now realize that this approach does not align with the current methodology. The updated approach is to return the closest index file if the requested file does not exist.

* Adding Tests

These tests will ensure serving and transformation of .html and .htm files when request explicitly includes filenames.

* Use only .html files

* Using endsWith instead of match

* Using meaningful variable/function names